+++ src/cmd.rs
@@ -4,17 +4,14 @@ use structopt::StructOpt;
 #[derive(StructOpt)]
 #[structopt(name = "dmenv", about = "The stupid virtualenv manager")]
 pub struct Command {
-    #[structopt(long = "python", help = "python binary",)]
+    #[structopt(long = "python", help = "python binary")]
     pub python_binary: Option<String>,
 
     // Those are mainly useful for tests, but you never know:
     #[structopt(long = "cfg-path", help = "path to the config file")]
     pub cfg_path: Option<String>,
 
-    #[structopt(
-        long = "cwd",
-        help = "path to use as the workining directory"
-    )]
+    #[structopt(long = "cwd", help = "path to use as the workining directory")]
     pub working_dir: Option<String>,
 
     #[structopt(subcommand)]
@@ -33,30 +30,24 @@ pub enum SubCommand {
     Init {
         #[structopt(long = "name", help = "Project name")]
         name: String,
-        #[structopt(
-            long = "version",
-            help = "Project version",
-            default_value = "0.1.0"
-        )]
+
+        #[structopt(long = "version", help = "Project version", default_value = "0.1.0")]
         version: String,
+
+        #[structopt(long = "author", help = "author")]
+        author: Option<String>,
     },
 
     #[structopt(name = "lock", about = "(Re)-generate requirements.lock")]
     Lock {},
 
-    #[structopt(
-        name = "run",
-        about = "Run the given binary from the virtualenv"
-    )]
+    #[structopt(name = "run", about = "Run the given binary from the virtualenv")]
     Run {
         #[structopt(name = "command")]
         cmd: Vec<String>,
     },
 
-    #[structopt(
-        name = "upgrade-pip",
-        about = "Upgrade pip in the virtualenv"
-    )]
+    #[structopt(name = "upgrade-pip", about = "Upgrade pip in the virtualenv")]
     UpgradePip {},
 
     #[structopt(name = "show", about = "Show path of the virtualenv")]