+++ tests/integration_tests.rs
@@ -18,8 +18,18 @@ fn init_generates_setup_py() {
     let tmp_dir = tempdir::TempDir::new("test-dmenv").expect("");
     let test_app = TestApp::new(tmp_dir.path().to_path_buf());
     test_app.remove_setup_py();
-    test_app.assert_run_ok(vec!["init", "--name", "foo"]);
-    test_app.assert_setup_py();
+    #[cfg_attr(rustfmt, rustfmt_skip)]
+    test_app.assert_run_ok(vec![
+        "init",
+        "--name", "foo",
+        "--version", "0.42",
+        "--author", "jane@corp.com",
+    ]);
+
+    let written = test_app.read_setup_py();
+    assert!(written.contains("foo"));
+    assert!(written.contains("0.42"));
+    assert!(written.contains("jane@corp.com"));
 }
 
 #[test]
@@ -56,7 +66,6 @@ fn install_workflow() {
     test_app.write_lock(&lock_contents);
     test_app.assert_run_ok(vec!["install"]);
 }
-
 
 #[test]
 fn run_without_args() {