+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # next
 
 * Add a `--author` option to `dmenv init`, used when generating the `setup.py` file.
-* `dmenv lock` now exits immediately if the lock file is missing. (#12)
+* Improve `dmenv lock`:
+  * Exit immediately if the lock file is missing. (#12)
+  * Always clean the previous virtualenv if it exists (#14)
+
 
 # 0.4.2